+++ frontend/src/@types/index.ts
@@ -1,0 +1,3 @@
+export { UserData } from "./user";
+export { Room } from "./room";
+export { Message } from "./message";
